+++ windows/video_outlet.h
@@ -1,3 +1,21 @@
+// This file is a part of dart_vlc (https://github.com/alexmercerind/dart_vlc)
+//
+// Copyright (C) 2021-2022 Hitesh Kumar Saini <saini123hitesh@gmail.com>
+//
+// This program is free software; you can redistribute it and/or
+// modify it under the terms of the GNU Lesser General Public
+// License as published by the Free Software Foundation; either
+// version 3 of the License, or (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+// Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with this program; if not, write to the Free Software Foundation,
+// Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
 #ifndef VIDEO_OUTLET_H_
 #define VIDEO_OUTLET_H_
 
@@ -13,7 +31,7 @@ class VideoOutlet {
 
   int64_t texture_id() const { return texture_id_; }
 
-  void OnVideo(uint8_t* buffer, int32_t width, int32_t height);
+  void MarkVideoFrameAvailable(uint8_t* buffer, int32_t width, int32_t height);
 
   ~VideoOutlet();